requirements.txt
# Flask Framework
Flask==3.0.0

# Security & Authentication
Werkzeug==3.0.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Production Server (Optional)
gunicorn==21.2.0

